+++ packages/registration_delivery/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.4-dev.2-uniqueId
+
+* Unique Beneficiary Id link to identifier
+* Updated digit_data_model package
+
 ## 1.0.4-dev.1-uniqueId
 
 * Added BeneficiaryInEligible status